+++ lua/lazyvim/plugins/extras/ui/mini-indentscope.lua
@@ -14,29 +14,53 @@ return {
     init = function()
       vim.api.nvim_create_autocmd("FileType", {
         pattern = {
-          "help",
+          "Trouble",
           "alpha",
           "dashboard",
-          "neo-tree",
-          "Trouble",
-          "trouble",
+          "fzf",
+          "help",
           "lazy",
           "mason",
+          "neo-tree",
           "notify",
+          "snacks_dashboard",
+          "snacks_notif",
+          "snacks_terminal",
+          "snacks_win",
           "toggleterm",
-          "lazyterm",
+          "trouble",
         },
         callback = function()
           vim.b.miniindentscope_disable = true
         end,
       })
+
+      vim.api.nvim_create_autocmd("User", {
+        pattern = "SnacksDashboardOpened",
+        callback = function(data)
+          vim.b[data.buf].miniindentscope_disable = true
+        end,
+      })
     end,
   },
+
+  -- disable inent-blankline scope when mini-indentscope is enabled
   {
     "lukas-reineke/indent-blankline.nvim",
+    optional = true,
     event = "LazyFile",
     opts = {
       scope = { enabled = false },
+    },
+  },
+
+  -- disable snacks scroll when mini-indentscope is enabled
+  {
+    "snacks.nvim",
+    opts = {
+      indent = {
+        scope = { enabled = false },
+      },
     },
   },
 }