+++ lua/lazyvim/util/lsp.lua
@@ -22,12 +22,13 @@ function M.get_clients(opts)
 end
 
 ---@param on_attach fun(client:vim.lsp.Client, buffer)
-function M.on_attach(on_attach)
+---@param name? string
+function M.on_attach(on_attach, name)
   return vim.api.nvim_create_autocmd("LspAttach", {
     callback = function(args)
       local buffer = args.buf ---@type number
       local client = vim.lsp.get_client_by_id(args.data.client_id)
-      if client then
+      if client and (not name or client.name == name) then
         return on_attach(client, buffer)
       end
     end,
@@ -43,12 +44,13 @@ function M.setup()
     ---@diagnostic disable-next-line: no-unknown
     local ret = register_capability(err, res, ctx)
     local client = vim.lsp.get_client_by_id(ctx.client_id)
-    local buffer = vim.api.nvim_get_current_buf()
     if client then
-      vim.api.nvim_exec_autocmds("User", {
-        pattern = "LspDynamicCapability",
-        data = { client_id = client.id, buffer = buffer },
-      })
+      for buffer in pairs(client.attached_buffers) do
+        vim.api.nvim_exec_autocmds("User", {
+          pattern = "LspDynamicCapability",
+          data = { client_id = client.id, buffer = buffer },
+        })
+      end
     end
     return ret
   end
@@ -58,6 +60,18 @@ end
 
 ---@param client vim.lsp.Client
 function M._check_methods(client, buffer)
+  -- don't trigger on invalid buffers
+  if not vim.api.nvim_buf_is_valid(buffer) then
+    return
+  end
+  -- don't trigger on non-listed buffers
+  if not vim.bo[buffer].buflisted then
+    return
+  end
+  -- don't trigger on nofile buffers
+  if vim.bo[buffer].buftype == "nofile" then
+    return
+  end
   for method, clients in pairs(M._supports_method) do
     clients[client] = clients[client] or {}
     if not clients[client][buffer] then
@@ -104,32 +118,24 @@ function M.on_supports_method(method, fn)
   })
 end
 
----@param from string
----@param to string
-function M.on_rename(from, to)
-  local clients = M.get_clients()
-  for _, client in ipairs(clients) do
-    if client.supports_method("workspace/willRenameFiles") then
-      ---@diagnostic disable-next-line: invisible
-      local resp = client.request_sync("workspace/willRenameFiles", {
-        files = {
-          {
-            oldUri = vim.uri_from_fname(from),
-            newUri = vim.uri_from_fname(to),
-          },
-        },
-      }, 1000, 0)
-      if resp and resp.result ~= nil then
-        vim.lsp.util.apply_workspace_edit(resp.result, client.offset_encoding)
-      end
-    end
-  end
-end
-
 ---@return _.lspconfig.options
 function M.get_config(server)
   local configs = require("lspconfig.configs")
   return rawget(configs, server)
+end
+
+---@return {default_config:lspconfig.Config}
+function M.get_raw_config(server)
+  local ok, ret = pcall(require, "lspconfig.configs." .. server)
+  if ok then
+    return ret
+  end
+  return require("lspconfig.server_configurations." .. server)
+end
+
+function M.is_enabled(server)
+  local c = M.get_config(server)
+  return c and c.enabled ~= false
 end
 
 ---@param server string
@@ -197,75 +203,37 @@ function M.format(opts)
   end
 end
 
----@alias LspWord {from:{[1]:number, [2]:number}, to:{[1]:number, [2]:number}, current?:boolean} 1-0 indexed
-M.words = {}
-M.words.ns = vim.api.nvim_create_namespace("vim_lsp_references")
-
----@param opts? {enabled?: boolean}
-function M.words.setup(opts)
-  opts = opts or {}
-  if not opts.enabled then
-    return
-  end
-  local handler = vim.lsp.handlers["textDocument/documentHighlight"]
-  vim.lsp.handlers["textDocument/documentHighlight"] = function(err, result, ctx, config)
-    if not vim.api.nvim_buf_is_loaded(ctx.bufnr) then
-      return
+M.action = setmetatable({}, {
+  __index = function(_, action)
+    return function()
+      vim.lsp.buf.code_action({
+        apply = true,
+        context = {
+          only = { action },
+          diagnostics = {},
+        },
+      })
     end
-    return handler(err, result, ctx, config)
-  end
+  end,
+})
 
-  M.on_supports_method("textDocument/documentHighlight", function(_, buf)
-    vim.api.nvim_create_autocmd({ "CursorHold", "CursorHoldI", "CursorMoved", "CursorMovedI" }, {
-      group = vim.api.nvim_create_augroup("lsp_word_" .. buf, { clear = true }),
-      buffer = buf,
-      callback = function(ev)
-        if not M.words.at() then
-          if ev.event:find("CursorMoved") then
-            vim.lsp.buf.clear_references()
-          else
-            vim.lsp.buf.document_highlight()
-          end
-        end
-      end,
+---@class LspCommand: lsp.ExecuteCommandParams
+---@field open? boolean
+---@field handler? lsp.Handler
+
+---@param opts LspCommand
+function M.execute(opts)
+  local params = {
+    command = opts.command,
+    arguments = opts.arguments,
+  }
+  if opts.open then
+    require("trouble").open({
+      mode = "lsp_command",
+      params = params,
     })
-  end)
-end
-
----@return LspWord[]
-function M.words.get()
-  local cursor = vim.api.nvim_win_get_cursor(0)
-  return vim.tbl_map(function(extmark)
-    local ret = {
-      from = { extmark[2] + 1, extmark[3] },
-      to = { extmark[4].end_row + 1, extmark[4].end_col },
-    }
-    if cursor[1] >= ret.from[1] and cursor[1] <= ret.to[1] and cursor[2] >= ret.from[2] and cursor[2] <= ret.to[2] then
-      ret.current = true
-    end
-    return ret
-  end, vim.api.nvim_buf_get_extmarks(0, M.words.ns, 0, -1, { details = true }))
-end
-
----@param words? LspWord[]
----@return LspWord?, number?
-function M.words.at(words)
-  for idx, word in ipairs(words or M.words.get()) do
-    if word.current then
-      return word, idx
-    end
-  end
-end
-
-function M.words.jump(count)
-  local words = M.words.get()
-  local _, idx = M.words.at(words)
-  if not idx then
-    return
-  end
-  local target = words[idx + count]
-  if target then
-    vim.api.nvim_win_set_cursor(0, target.from)
+  else
+    return vim.lsp.buf_request(0, "workspace/executeCommand", params, opts.handler)
   end
 end