+++ lua/lazyvim/util/lualine.lua
@@ -1,13 +1,37 @@
 ---@class lazyvim.util.lualine
 local M = {}
 
+---@param icon string
+---@param status fun(): nil|"ok"|"error"|"pending"
+function M.status(icon, status)
+  local colors = {
+    ok = "Special",
+    error = "DiagnosticError",
+    pending = "DiagnosticWarn",
+  }
+  return {
+    function()
+      return icon
+    end,
+    cond = function()
+      return status() ~= nil
+    end,
+    color = function()
+      return { fg = Snacks.util.color(colors[status()] or colors.ok) }
+    end,
+  }
+end
+
+---@param name string
+---@param icon? string
 function M.cmp_source(name, icon)
+  icon = icon or LazyVim.config.icons.kinds[name:sub(1, 1):upper() .. name:sub(2)]
   local started = false
-  local function status()
+  return M.status(icon, function()
     if not package.loaded["cmp"] then
       return
     end
-    for _, s in ipairs(require("cmp").core.sources) do
+    for _, s in ipairs(require("cmp").core.sources or {}) do
       if s.name == name then
         if s.source:is_available() then
           started = true
@@ -20,25 +44,7 @@ function M.cmp_source(name, icon)
         return "ok"
       end
     end
-  end
-
-  local colors = {
-    ok = LazyVim.ui.fg("Special"),
-    error = LazyVim.ui.fg("DiagnosticError"),
-    pending = LazyVim.ui.fg("DiagnosticWarn"),
-  }
-
-  return {
-    function()
-      return icon or require("lazyvim.config").icons.kinds[name:sub(1, 1):upper() .. name:sub(2)]
-    end,
-    cond = function()
-      return status() ~= nil
-    end,
-    color = function()
-      return colors[status()] or colors.ok
-    end,
-  }
+  end)
 end
 
 ---@param component any
@@ -72,7 +78,7 @@ function M.format(component, text, hl_group)
   return component:format_hl(lualine_hl_group) .. text .. component:get_default_hl()
 end
 
----@param opts? {relative: "cwd"|"root", modified_hl: string?, directory_hl: string?, filename_hl: string?}
+---@param opts? {relative: "cwd"|"root", modified_hl: string?, directory_hl: string?, filename_hl: string?, modified_sign: string?, readonly_icon: string?, length: number?}
 function M.pretty_path(opts)
   opts = vim.tbl_extend("force", {
     relative = "cwd",
@@ -80,6 +86,7 @@ function M.pretty_path(opts)
     directory_hl = "",
     filename_hl = "Bold",
     modified_sign = "",
+    readonly_icon = " 󰌾 ",
     length = 3,
   }, opts or {})
 
@@ -90,12 +97,13 @@ function M.pretty_path(opts)
       return ""
     end
 
+    path = LazyVim.norm(path)
     local root = LazyVim.root.get({ normalize = true })
     local cwd = LazyVim.root.cwd()
 
     if opts.relative == "cwd" and path:find(cwd, 1, true) == 1 then
       path = path:sub(#cwd + 2)
-    else
+    elseif path:find(root, 1, true) == 1 then
       path = path:sub(#root + 2)
     end
 
@@ -105,7 +113,7 @@ function M.pretty_path(opts)
     if opts.length == 0 then
       parts = parts
     elseif #parts > opts.length then
-      parts = { parts[1], "…", table.concat({ unpack(parts, #parts - opts.length + 2, #parts) }, sep) }
+      parts = { parts[1], "…", unpack(parts, #parts - opts.length + 2, #parts) }
     end
 
     if opts.modified_hl and vim.bo.modified then
@@ -120,7 +128,12 @@ function M.pretty_path(opts)
       dir = table.concat({ unpack(parts, 1, #parts - 1) }, sep)
       dir = M.format(self, dir .. sep, opts.directory_hl)
     end
-    return dir .. parts[#parts]
+
+    local readonly = ""
+    if vim.bo.readonly then
+      readonly = M.format(self, opts.readonly_icon, opts.modified_hl)
+    end
+    return dir .. parts[#parts] .. readonly
   end
 end
 
@@ -132,7 +145,9 @@ function M.root_dir(opts)
     parent = true,
     other = true,
     icon = "󱉭 ",
-    color = LazyVim.ui.fg("Special"),
+    color = function()
+      return { fg = Snacks.util.color("Special") }
+    end,
   }, opts or {})
 
   local function get()